Removed kirin.dependencies array.
Previously, kirin-build was using an array of dependencies in the kirinBlock. This was causing too much confusion in how to add dependencies (app devs had to add it in two places in the same file).

Dependencies are now being discovered via the package's own dependencies block.

This means that we can now use

 % npm install foo --save

But it does mean that kirin-build and progenitor are now devDependencies; installation of kirin is will now:

 % npm --dev install kirin

TODO change in documentation.